+++ src/main/java/com/mlt/api/apioperations/repository/OperationTypeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface OperationTypeRepository extends JpaRepository<OperationTypeModel, Long> {
+public interface OperationTypeRepository extends JpaRepository<OperationTypeModel, Integer> {
 
     Optional<OperationTypeModel> findByDescriptionIgnoreCase(String description);
